requirements.txt
django
djangorestframework
django-rest-knox
django-rest-passwordreset
django-cors-headers